+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 5.2.0 - 2020-06-01
+
 ### Added
 - A new method, `getPayment`, added to `XRPClient` for retrieving payment transactions by hash.
 - `XrplNetwork` provides the functionality of `XRPLNetwork` under an idiomatic cased name.